move some functions to other file
